+++ internal/cmd/genlangdocs/main.go
@@ -51,6 +51,10 @@ func main() {
 	langs := langConfigs(make([]langConfig, 0, len(scanner.LanguagesConfig)))
 
 	for lang, config := range scanner.LanguagesConfig {
+		if lang == scanner.GenericLanguage {
+			continue
+		}
+
 		info, err := enry.GetLanguageInfo(lang)
 		if err != nil {
 			panic(err)
@@ -69,7 +73,7 @@ func main() {
 	fmt.Println("")
 	fmt.Println("# Supported Languages")
 	fmt.Println("")
-	fmt.Printf("%d languages are currently supported.\n", len(scanner.LanguagesConfig))
+	fmt.Printf("%d languages are currently supported.\n", len(langs))
 	fmt.Println("")
 
 	fmt.Println("| Language | Files | Supported comments |")